Change html.js to html.jsx
